@@ -4,44 +4,48 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
-  bloc: ^8.0.1
-  cached_network_image: ^3.2.0
-  cupertino_icons: ^1.0.4
+  bloc: ^8.1.0
+  cached_network_image: ^3.2.3
+  cupertino_icons: ^1.0.5
   dartz: ^0.10.1
-  dio: ^4.0.4
-  equatable: ^2.0.3
+  dio: ^4.0.6
+  equatable: ^2.0.5
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.0.0
+  flutter_bloc: ^8.1.1
   flutter_localizations:
     sdk: flutter
-  flutter_svg: ^1.0.0
-  functional_enum_annotation: ^1.2.2
+  flutter_svg: ^1.1.6
+  functional_enum_annotation: ^1.2.3
   get_it: ^7.2.0
-  go_router: ^2.5.3
-  hive: ^2.0.4
+  go_router: ^6.0.0
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
-  injectable: ^1.5.0
+  injectable: ^2.1.0
   intl: ^0.17.0
-  json_annotation: ^4.4.0
+  json_annotation: ^4.7.0
+  networkx: ^0.1.0
   pretty_dio_logger: ^1.1.1
-  stream_transform: ^2.0.0
-  url_launcher: ^6.0.17
+  stream_transform: ^2.1.0
+  url_launcher: ^6.1.7
 
 dev_dependencies:
-  bloc_test: ^9.0.1
-  build_runner: ^2.1.5
-  flutter_launcher_icons: ^0.9.2
+  bloc_test: ^9.1.0
+  build_runner: ^2.3.3
+  flutter_launcher_icons: ^0.11.0
   flutter_test:
     sdk: flutter
-  functional_enum: ^1.1.4
-  hive_generator: ^1.1.1
-  injectable_generator: ^1.5.2
-  json_serializable: ^6.1.1
-  mocktail: ^0.2.0
+  functional_enum: ^1.1.6
+  hive_generator: ^2.0.0
+  injectable_generator: ^2.1.3
+  json_serializable: ^6.5.4
+  mocktail: ^0.3.0
+
+dependency_overrides:
+  test_api: ^0.4.17
 
 flutter:
   uses-material-design: true
